# Caches expensive cudnn graph validation