@parthmannan 2024-04-24
Verified that the PR fixes #241 

Updated timings -

LLaMa2 7B FSDP ZeRO2 performance
Microbatch=1, Global batch=8 on 8xH100

```
thunder_inductor: 404.71 ms/iter
thunder_inductor_cudnn: 384.99 ms/iter
```
Launch overhead is significantly reduced. Down to 400us behind kernel launches in some cases.
Thanks @vedaanta-nvidia, this looks great!